refactor(webapp): drop scoops preview-minter shim after shell move
Point tests and mocks at shell/preview-minter and remove the unused scoops
re-export so the production dead-code gate stays green.
