--- packages/webapp/src/scoops/preview-minter.ts
@@ -1,2 +1,0 @@
-/** @deprecated Import from `shell/preview-minter.js` — relocated to fix layer back-edge. */
-export * from '../shell/preview-minter.js';
